fea: refactor types Fixes #51
