+++ src/libs/type/index.ts
@@ -1,0 +1,5 @@
+// map and lists
+export { iconList } from './lib/icon';
+
+// types
+export type { TIconList } from './lib/icon';
